feat: add --force and --dry-run options

+++ src/index.ts
@@ -4,7 +4,7 @@ import * as process from "process";
 import * as path from "path";
 import * as fs from "fs";
 import * as cp from "child_process";
-import { keyBy } from "lodash";
+import { keyBy, pick } from "lodash";
 import * as sv from "semver";
 import { unparse } from "./argv";
 
@@ -29,12 +29,21 @@ export type LockfileObject = {
 
 export type IFixOptions = {
   cwd?: string;
+  dryRun?: boolean;
+  force?: boolean;
+  groups?: string | string[];
+  level?: string;
+  verbose?: boolean;
 };
 
 export const run = (opts: IFixOptions): void => {
   const { cwd = process.cwd() } = opts;
   const yarnlockPath = path.resolve(cwd, "yarn.lock");
-  const yarnAuditCmd = unparse({ ...opts, json: true, _: [] }, { command: "yarn audit" }).join(" ");
+  const yarnAuditAllowedOpts = ["cwd", "groups", "level", "verbose"];
+  const yarnAuditCmd = unparse(
+    { ...pick(opts, yarnAuditAllowedOpts), json: true, _: [] },
+    { command: "yarn audit" }
+  ).join(" ");
   const spawnOptions = {
     shell: true,
     maxBuffer: 128 * 1024 * 1024,
@@ -102,7 +111,7 @@ export const run = (opts: IFixOptions): void => {
         );
         continue;
       }
-      if (!sv.satisfies(fix, desiredRange)) {
+      if (!sv.satisfies(fix, desiredRange) && !opts.force) {
         console.error(
           "Cant find patched version that satisfies",
           depSpec,
@@ -119,7 +128,9 @@ export const run = (opts: IFixOptions): void => {
     }
   }
 
-  fs.writeFileSync(yarnlockPath, lf.stringify(lockfile));
+  if (!opts.dryRun) {
+    fs.writeFileSync(yarnlockPath, lf.stringify(lockfile));
+  }
 
   console.log("Installing upgrades:", upgradeVersions.join(", "));
   cp.spawnSync(`yarn install --update-checksums --cwd ${cwd}`, spawnOptions);
